#142 moved test for items to a new file
